+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/management/ui/menu/UIMenuItemSetting.java
@@ -16,16 +16,28 @@
  *
  */
 
-package org.apache.skywalking.oap.server.core;
+package org.apache.skywalking.oap.server.core.management.ui.menu;
 
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
+import lombok.Data;
 
-public class CoreModuleTest {
-    @Test
-    public void testOpenServiceList() {
-        CoreModule coreModule = new CoreModule();
+import java.util.List;
 
-        Assertions.assertEquals(41, coreModule.services().length);
-    }
+@Data
+public class UIMenuItemSetting {
+    /**
+     * Title of the menu item.
+     */
+    private String title;
+    /**
+     * Icon name of the menu item, it should only exist in the top level of menu.
+     */
+    private String icon;
+    /**
+     * The layer name of the menu item.
+     */
+    private String layer;
+    /**
+     * Sub menus of current item.
+     */
+    private List<UIMenuItemSetting> menus;
 }