+++ docs/en/changes/changes.md
@@ -28,6 +28,7 @@
 * `BanyanDBMetricsDAO` handles `storeIDTag` in `multiGet` for `BanyanDBModelExtension`.
 * Fix endpoint grouping-related logic and enhance the performance of PatternTree retrieval. 
 * Fix metric session cache saving after batch insert when using `mysql-connector-java`
+* Support dynamic UI menu query.
 
 #### UI